pyproject.toml
[project]
name = "docsvision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "numpy>=2.4.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]
